Split out new components to new files
