+++ src/types/item.ts
@@ -1,0 +1,6 @@
+export interface Item {
+  id: string;
+  content: string;
+  created: string;
+}
+
